requirements.txt
Flask
Werkzeug
tensorflow
opencv-python
Pillow